requirements.txt
streamlit
pandas
numpy
requests
scikit-learn
tensorflow-cpu
geopy
pytz
folium
streamlit-folium
